requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
python-decouple
django_filter
drf_yasg